+++ librum/src/i_stream/mod.rs
@@ -19,6 +19,7 @@
 // er General Public License along with Librum. If
 // not, see <https://www.gnu.org/licenses/>.
 
+use core::ptr::copy_nonoverlapping;
 use core::slice;
 
 /// Byte stream suitable for input.
@@ -37,11 +38,13 @@ impl<'a> IStream<'a> {
 
 	/// Reads bytes from the stream.
 	///
+	/// This method may be preferred over [`read_into`](Self::read_into) if the read data isn't directly needed, e.g. if an iterator is applied anyway to map the data.
+	///
 	/// # Panics
 	///
 	/// If the requested amount of bytes could not exactly be read, then this method will panic.
 	#[inline]
-	pub fn read(&mut self, count: usize) -> &[u8] {
+	pub fn read(&mut self, count: usize) -> &'a [u8] {
 		let remaining = self.buf.len() - self.pos;
 
 		assert!(
@@ -60,6 +63,35 @@ impl<'a> IStream<'a> {
 		self.pos += count;
 
 		data
+	}
+
+	/// Reads bytes from the stream into a predefined buffer.
+	///
+	/// This method may be preferred over [`read`](Self::read) if the read data **is** directly needed, e.g. if all required transformations can be done in-place.
+	///
+	/// # Panics
+	///
+	/// If the provided buffer could not be completely filled, then this method will panic.
+	#[inline]
+	pub fn read_into(&mut self, buf: &mut [u8]) {
+		let count     = buf.len();
+		let remaining = self.buf.len() - self.pos;
+
+		assert!(
+			remaining >= count,
+			"cannot read ({count}) bytes at ({}) from stream with capacity of ({})",
+			self.pos,
+			self.buf.len(),
+		);
+
+		unsafe {
+			let src = self.buf.as_ptr().add(self.pos);
+			let dst = buf.as_mut_ptr();
+
+			copy_nonoverlapping(src, dst, count);
+		}
+
+		self.pos += count;
 	}
 
 	/// Closes the stream.